requirements.txt
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
uWSGI==2.0.18
Werkzeug==1.0.1
WTForms==2.3.1
uwsgi
